+++ R/api-dsl.R
@@ -114,6 +114,7 @@ map_to_grid <- function(transform, target, row, plan, graph) {
   new_targets <- new_targets(
     target, grid, cols = sub_cols, id = dsl_id(transform)
   )
+  grid$.id_chr <- sprintf("\"%s\"", new_targets)
   out <- data.frame(target = new_targets, stringsAsFactors = FALSE)
   for (col in setdiff(old_cols(plan), c("target", "transform"))) {
     if (is.language(row[[col]][[1]])) {